chore: UserList와 UserItem을 분류해 불필요한 리렌더링 방지
